README, Language Reference, Examples
If you're brave, this gives you everything you need to get started with Fictive. It's buggy, untested, and a gigantic mess, but you *could* start using it. If you don't mind things breaking in future versions.

But I *think* this approach to writing games is kinda fun. It's fun for me. But maybe I'm weird.