Multifile Loading Support
Added a dedicated loader, which can handle two cases: the legacy "read
from a single file" case, and a "read from a directory containing a
manifest.yaml" case. The latter makes it easier to work with multiple
files.

NB: I initially spent some time trying to do this with `!include`
directives in PyYAML, but those don't support anchors neatly, and
anchors are an easy way to organize your game. So the manifest file is a
compromise, that still lets you include multiple files.

The down side is that it's going to be a lot harder to track down errors
in your YAML, since what the YAML parser gets is the MERGED files, not
the individual files. That's a problem for future me. It's not like you
get good errors right now anyway.