First UI Pass
This is a first pass at having a UI, and it changes a LOT of things.

We adopted Textual as the UI. There are reasons, but mostly I started out
trying Curses and it sucked.

Because we have a real UI, the game loop changed. It now explicitly
returns the current state (which contains substates) and transients-
reverted states are transients. That lets us output messages.

That cleaned up the game loop a fair bit.

Added:
* A Loading Screen (to pick games from)
* A Game Screen (to play)

That all exists inside the textual app.

To support the loading screen, I also changed the manifest format. It now
lets you specify a title and a slug, and then your list of files.