Changed the State handling to only EVER transition one state
Previous behavior: substates and states could both transition.
Current behavior: it tries a substate transition, then a state transition,
then a global transition. It transitions on the first one that works.

The logic behind this is that a State may want to provide a "Help" or "Error"
transition when the user submits an input that doesn't trigger a transition,
but we definitely don't want that to happen if a substate transitioned.