Debugging Support
Adds a Debug flag to the launcher, which controls a "Peek Statebag" menu
option. That will add a pane to the view which shows the current status
of the statebag, helpful for debugging.

There needs to be more debugging functionality to really make this work.
Like I'd like another pane that shows the transitions and their
conditions. That's a more ambitious debugger.