Tutorial Game
This creates a tutorial game, which will teach designers how to use Fictive.

It's both instructional, and a basic test for my core engine. Is leverages all
the features of the engine, uncovered a bunch of bugs in the process, and now
lets us turn this into a tool people could actually *use*, which is important.