Makes Function Lookup Case Insensitive
Look, there's no reason for our engine to be picky about this. I don't
CARE what capitalization designers use, because it's not like they're
going to be writing custom triggers as part of game development. (Yet,
or for the forseeable, but making this extensible is probably a good
idea, but even then- we can keep it case insensitive).