Removed PyYAML, Switched to Ruamel.Yaml
I've gone back and forth on this a few times, but PyYAML doesn't seem
to have YAML1.2 support, and the result is that "on" and "off" behave
in the stupid YAML fashion. Ruamel is YAML1.2, and this fixes the issues
I was having with the example game with only minor changes to my code,
instead of trying to rewrite the entire goddamn type coercion and likely
running into some sort of mistake there, because I hate type coercion.

I do suspect this will still bite me at some point with boolean values, but we'll see.