Mildly Better Error Handling
Now, when we have a broken YAML input, it won't crash the application.
We still don't provide a good report so the developer can debug their
code. That's something I have to think through more, how that's going to
work.

But at least the player gets a little notification and the app doesn't
crash.